Conflicts:
	dante5-stock/build.gradle
	dante5-stock/src/main/java/com/shangpin/iog/dante5/stock/GrabStockImp.java

+++ iog-web-container/src/main/java/com/shangpin/iog/webcontainer/front/util/Excel2DTO.java
@@ -76,7 +76,7 @@ public class Excel2DTO {
 			strCell = xssfCell.toString();
 			break;
 		case XSSFCell.CELL_TYPE_NUMERIC:
-			strCell = String.valueOf((int)xssfCell.getNumericCellValue());
+			strCell = String.valueOf((long)xssfCell.getNumericCellValue());
 			break;
 		case XSSFCell.CELL_TYPE_BOOLEAN:
 			strCell = String.valueOf(xssfCell.getStringCellValue());
